+++ Runtime/HttpResponses/HttpResponse[T,E].cs
@@ -14,7 +14,7 @@ namespace ApiClient.Runtime.HttpResponses
     /// </summary>
     /// <typeparam name="T"></typeparam>
     /// <typeparam name="E"></typeparam>
-    public class HttpResponse<T, E> : IHttpResponse, IHttpResponseStatusCode, IHttpResponseBody, ICachedHttpResponse
+    public class HttpResponse<T, E> : IHttpResponse, IHttpResponseStatusCode, IHttpResponseBody, IHttpResponseTiming, ICachedHttpResponse
     {
         public HttpResponse(T content, E error, HttpResponseHeaders headers, HttpContentHeaders contentHeaders, string body, HttpRequestMessage request, HttpStatusCode statusCode)
         {
@@ -26,6 +26,7 @@ namespace ApiClient.Runtime.HttpResponses
             StatusCode = statusCode;
             ContentHeaders = contentHeaders.ToHeadersDictionary();
             Body = body;
+            TimingInfo = new TimingInfo();
         }
 
         /// <summary>
@@ -66,6 +67,8 @@ namespace ApiClient.Runtime.HttpResponses
         /// </summary>
         /// <value></value>
         public string Body { get; private set; }
+
+        public TimingInfo TimingInfo { get; set; }
 
         bool ICachedHttpResponse.IsFromCache { get; set; }