+++ Runtime/HttpResponses/NetworkErrorHttpResponse.cs
@@ -4,7 +4,7 @@ using System.Net.Http;
 
 namespace ApiClient.Runtime.HttpResponses
 {
-    public class NetworkErrorHttpResponse : IHttpResponse
+    public class NetworkErrorHttpResponse : IHttpResponse, IHttpResponseTiming
     {
         public bool IsClientError => false;
         public bool IsServerError => false;
@@ -16,6 +16,7 @@ namespace ApiClient.Runtime.HttpResponses
         public Uri RequestUri { get; private set; }
         public Dictionary<string, string> Headers { get; private set; } = null;
         public Dictionary<string, string> ContentHeaders { get; private set; } = null;
+        public TimingInfo TimingInfo { get; set; }
 
         public string Message { get; }
 
@@ -24,12 +25,14 @@ namespace ApiClient.Runtime.HttpResponses
             Message = message;
             RequestMethod = request.Method;
             RequestUri = request.RequestUri;
+            TimingInfo = new TimingInfo();
         }
 
         public NetworkErrorHttpResponse(string message, Uri requestUri)
         {
             Message = message;
             RequestUri = requestUri;
+            TimingInfo = new TimingInfo();
         }
     }
 }